+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # SRBuild Changelog
 Dates are in YYYY-MM-DD format.
 
+## v0.2.4 ()
+- Fixed an issue with duplicate libraries when linking.
+
 ## v0.2.3 (2019-05-04)
 - Added `test` to `cli()` and `Project` to provide a convenient test runner.